refactor: organise datastore and remote data files
